personal financial calculater

+++ python_introduction/finance_calculator.py
@@ -1,9 +1,9 @@
-monthly_income = float(input("Enter your monthly income:"))
-total_monthly_expenses = float(input("Enter your total monthly expenses:"))
+monthly_income = int(input("Enter your monthly income:"))
+total_monthly_expenses = int(input("Enter your total monthly expenses:"))
 
-monthly_savings = monthly_income - total_monthly_expenses
+monthly_savings = int(monthly_income - total_monthly_expenses)
 
-Projected_Savings = (monthly_savings * 12 + (monthly_savings * 12 * 0.05))
+Projected_Savings = int(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
 
 print(f"Your monthly savings are {monthly_savings}.")
 print(f"Projected savings after one year, with interest, is {Projected_Savings}.")
